remove breadcrumb thunks from controller-action files
https://coveord.atlassian.net/browse/KIT-316
