+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-controller-actions.ts
@@ -19,8 +19,7 @@ import {
   facetIdDefinition,
 } from '../generic/facet-actions-validation';
 import {validateCategoryFacetValue} from './category-facet-validate-payload';
-import {ArrayValue, NumberValue} from '@coveo/bueno';
-import {logCategoryFacetBreadcrumb} from './category-facet-set-analytics-actions';
+import {NumberValue} from '@coveo/bueno';
 
 /**
  * Toggles the facet value and then executes a search with the appropriate analytics tag.
@@ -81,39 +80,5 @@ export const executeDeselectAllCategoryFacetValues = createAsyncThunk<
     dispatch(updateCategoryFacetNumberOfValues({facetId, numberOfValues}));
     dispatch(updateFacetOptions({freezeFacetOrder: true}));
     dispatch(executeSearch(logFacetClearAll(facetId)));
-  }
-);
-
-/**
- * Selects the breadcrumb category facet value and then executes a search with the appropriate analytics tag.
- * @param facetId (string) The unique identifier of the facet (e.g., `"1"`).
- * @param numberOfValues (number) The number of category facet values to show after deselecting.
- * @param path (string[]) The currently selected path.
- */
-export const executeSelectCategoryFacetBreadcrumb = createAsyncThunk<
-  void,
-  {facetId: string; numberOfValues: number; path: string[]},
-  AsyncThunkSearchOptions<ConfigurationSection & CategoryFacetSection>
->(
-  'dateFacet/executeSelectCategoryFacetBreadcrumb',
-  ({facetId, numberOfValues, path}, {dispatch, extra: {validatePayload}}) => {
-    validatePayload(
-      {facetId, numberOfValues, path},
-      {
-        facetId: facetIdDefinition,
-        numberOfValues: new NumberValue({required: true}),
-        path: new ArrayValue({required: true, each: requiredNonEmptyString}),
-      }
-    );
-    dispatch(deselectAllCategoryFacetValues(facetId));
-    dispatch(updateCategoryFacetNumberOfValues({facetId, numberOfValues}));
-    dispatch(
-      executeSearch(
-        logCategoryFacetBreadcrumb({
-          categoryFacetPath: path,
-          categoryFacetId: facetId,
-        })
-      )
-    );
   }
 );
