+++ packages/headless/src/features/facets/facet-set/facet-set-controller-actions.ts
@@ -8,14 +8,10 @@ import {
 import {getAnalyticsActionForToggleFacetSelect} from './facet-set-utils';
 import {updateFacetOptions} from '../../facet-options/facet-options-actions';
 import {executeSearch} from '../../search/search-actions';
-import {
-  toggleSelectFacetValue,
-  selectFacetBreadcrumb,
-} from './facet-set-actions';
+import {toggleSelectFacetValue} from './facet-set-actions';
 import {facetIdDefinition} from '../generic/facet-actions-validation';
 import {RecordValue} from '@coveo/bueno';
 import {facetValueDefinition} from './facet-set-validate-payload';
-import {logFacetBreadcrumb} from './facet-set-analytics-actions';
 
 const definition = {
   facetId: facetIdDefinition,
@@ -44,31 +40,6 @@ export const executeToggleFacetSelect = createAsyncThunk<
     validatePayload({facetId, selection}, definition);
     dispatch(toggleSelectFacetValue({facetId, selection}));
     dispatch(updateFacetOptions({freezeFacetOrder: true}));
-    dispatch(executeSearch(analyticsAction));
-  }
-);
-
-/**
- * Selects the breadcrumb facet value and then executes a search with the appropriate analytics tag.
- * @param facetId (string) The unique identifier of the facet (e.g., `"1"`).
- * @param selection (FacetValue) The target facet value.
- */
-export const executeSelectFacetBreadcrumb = createAsyncThunk<
-  void,
-  {
-    facetId: string;
-    selection: FacetValue;
-  },
-  AsyncThunkSearchOptions<FacetSection & ConfigurationSection>
->(
-  'facet/executeSelectFacetBreadcrumb',
-  ({facetId, selection}, {dispatch, extra: {validatePayload}}) => {
-    const analyticsAction = logFacetBreadcrumb({
-      facetId: facetId,
-      facetValue: selection.value,
-    });
-    validatePayload({facetId, selection}, definition);
-    dispatch(selectFacetBreadcrumb({facetId, selection}));
     dispatch(executeSearch(analyticsAction));
   }
 );
